document props
